Extend memory offsets to 64 bits

+++ compiler/src/main/java/org/qbicc/interpreter/VmArray.java
@@ -8,7 +8,7 @@ import org.qbicc.type.ArrayObjectType;
 public interface VmArray extends VmObject {
     int getLength();
 
-    int getArrayElementOffset(int index);
+    long getArrayElementOffset(int index);
 
     ArrayObjectType getObjectType();
